% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{children}
\alias{children}
\title{Immediate descendants of all nodes}
\usage{
children(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\value{
A list holding the node indices of the immediate descendants of
each node.
}
\description{
Immediate descendants of all nodes
}
