% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroevolution.R
\docType{package}
\name{macroevolution}
\alias{macroevolution}
\title{macroevolution: a package for comparative analysis of biological diversity}
\section{Tree I/O functions}{

read.newick
write.newick
}

\section{Tree info functions}{

ages
ancestors
brlens
children
descendants
tips
Nnode
Ntip
root
tiplabels
tree.isbinary
tree.isultrametric
}

\section{Tree manipulation functions}{

extract.clade
keep.tip
drop.tip
rotate.node
tree.duplicate
tree.ladderize
}

\section{Tree visualization functions}{

plot.tree
}

\section{Functions for models of character evolution}{

make.mk.dmm
make.mk.dmm.from.sample
read.mk.dmm
make.rcm.dmm
make.rcm.dmm.from.sample
read.rcm.dmm
make.mk.epoch
make.ratematrix.mk.epoch
make.asr.mk.epoch
}

