% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_epoch.R
\name{make.asr.mk.epoch}
\alias{make.asr.mk.epoch}
\title{Marginal ancestral state reconstruction}
\usage{
make.asr.mk.epoch(lik)
}
\arguments{
\item{lik}{A likelihood function returned by \code{make.mk.epoch}}
}
\value{
A function that computes marginal ancestral state estimates for
 the internal nodes of a phylogeny given a set of transition rate
 parameters.
}
\description{
Marginal ancestral state reconstruction
}
\author{
Michael C. Grundler
}
