% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_dmm.R
\name{make.mk.dmm.from.sample}
\alias{make.mk.dmm.from.sample}
\title{Create a summary object from a single posterior sample}
\usage{
make.mk.dmm.from.sample(i, output.file)
}
\arguments{
\item{i}{The index of the posterior sample to summarize}

\item{output.file}{The name of the output.file}
}
\value{
A list with following structure
\describe{
 \item{asr}{Marginal ancestral state probabilities.}
 \item{dens.mean}{Posterior mean estimate of each multinomial in the sample.}
 \item{dens.map}{Max a posteriori estimate of each multinomial in the sample.}
 \item{dens.sampl}{Function to sample from the posterior distribution of each
  multinomial in this sample.}
}
}
\description{
Create a summary object from a single posterior sample
}
