% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_epoch.R
\name{make.ratematrix.mk.epoch}
\alias{make.ratematrix.mk.epoch}
\title{Initialize epoch transition rate matrices}
\usage{
make.ratematrix.mk.epoch(layout, covarion = FALSE)
}
\arguments{
\item{layout}{An array with dim=c(nstate, nstate, nepoch). Each matrix in
the array specifies the structure of the transition rate matrix for the
corresponding epoch. A value of 0 means that the given transition rate is
constrained to be 0. Any other value corresponds to the index of a transition
rate parameter. Therefore, different transitions can be constrained
to have the same rate by constraining them to share the same parameter.}

\item{covarion}{A boolean value indicating whether or not \code{layout}
is setup for a covarion-like model.}
}
\value{
An object that inherits class \code{ratematrix.mk.epoch} that can
 be used with \code{make.mk.epoch} to construct a likelihood function.
}
\description{
Initialize epoch transition rate matrices
}
\author{
Michael C. Grundler
}
