% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_dmm.R
\name{make.rcm.dmm.from.sample}
\alias{make.rcm.dmm.from.sample}
\title{Create a summary object from a single posterior sample}
\usage{
make.rcm.dmm.from.sample(i, output.file)
}
\arguments{
\item{i}{The index of the posterior sample to summarize}

\item{output.file}{The name of the output.file}
}
\value{
A list with following structure
\describe{
 \item{asr}{Marginal ancestral state probabilities.}
 \item{smap}{A function to perform stochastic character mapping.}
 \item{dens.map}{Max a posteriori estimate of each multinomial in the sample.}
 \item{dens.mean}{Posterior mean estimate of each multinomial in the sample.}
 \item{tip.state}{Terminal node state assignments. These align with the row
 indices of \code{dens.map} and \code{dens.mean}.}
}
}
\description{
Create a summary object from a single posterior sample
}
