% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_norm.R
\name{make.rcm.norm.from.sample}
\alias{make.rcm.norm.from.sample}
\title{Create a summary object from a single posterior sample}
\usage{
make.rcm.norm.from.sample(i, output.file)
}
\arguments{
\item{i}{The index of the posterior sample to summarize}

\item{output.file}{The name of the output.file}
}
\value{
A list with following structure
\describe{
 \item{asr}{Marginal ancestral state probabilities.}
 \item{smap}{A function to perform stochastic character mapping.}
 \item{tip.state}{Terminal node state assignments. These align with the row
 indices of \code{dens.map} and \code{dens.mean}.}
 \item{niv.pars}{Parameters of the posterior distribution for each state.}
 \item{sample}{Function to sample from each state's posterior.}
}
}
\description{
Create a summary object from a single posterior sample
}
