% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_norm.R
\name{make.rcm.norm}
\alias{make.rcm.norm}
\title{Normal-inverse Wishart Markov process model (random cluster model)}
\usage{
make.rcm.norm(
  phy,
  x,
  r,
  stateid.init,
  rate.init,
  kappa.init,
  nu.init,
  mu.init,
  lambda.init,
  integrate.brlen = FALSE,
  species.col = 1L,
  specimen.col = 0L
)
}
\arguments{
\item{phy}{An object of class \code{tree}.}

\item{x}{A \code{data.frame} containing continuous character data. Columns
correspond to characters and rows to species or specimens.}

\item{r}{The number of states in the prior model.}

\item{stateid.init}{An optional initial partition of terminal nodes into
states. If omitted all terminals are placed in a single cluster.}

\item{rate.init}{An optional initial value for the rate of the Poisson process
prior on the evolution of states. If omitted defaults to a random value.}

\item{kappa.init}{An optional initial value for the kappa hyperparameter of
the NIV prior on the normal distributions corresponding to each cluster.
If omitted defaults to 1. This is never updated during the MCMC.}

\item{nu.init}{An optional initial value for the nu hyperparameter of
the NIV prior on the normal distributions corresponding to each cluster.
If omitted defaults to 1. This is never updated during the MCMC.}

\item{mu.init}{An optional initial value for the mean vector hyperparameter of
the NIV prior on the normal distributions corresponding to each cluster.
If omitted defaults to the global mean vector. This is never updated during the MCMC.}

\item{lambda.init}{An optional initial value for the covariance matrix hyperparameter of
the NIV prior on the normal distributions corresponding to each cluster.
If omitted defaults to a diagonal matrix of sample variances.
This is never updated during the MCMC.}

\item{integrate.brlen}{An optional boolean value indicating whether or not to
assume branches have i.i.d. rates drawn from a Gamma prior. If TRUE, the branch
rates are integrated out of the likelihood function such that a single set of
transition probabilities applies to all branches. The default is FALSE, in
which case all branches share a single rate.}

\item{species.col}{The column index containing the species labels. The
default is 1.}

\item{specimen.col}{The column index containing the specimen labels. The
default is 0, meaning that character values correspond to species means.}
}
\value{
A function to perform MCMC inference that takes the following arguments
\describe{
\item{niter}{The number of iterations that the MCMC will run for. Default is 1000.}
\item{thin}{The number of iterations between each recorded sample. Default is 1.}
\item{tune.rate}{Tuning parameter for the rate. Should be between 0 and 1. Default is 0.1}
\item{update.node}{Weight for proposing node state updates. Default is 1.}
\item{update.rate}{Weight for proposing updates to the rate. Default is 1.}
\item{output.file}{Name of file for recording the state of Markov chain. Default is "mcmc.out".}
\item{output.mode}{One of either "wb" or "ab", indicating whether the output file should be opened
for writing or appending. Default is "wb".}
}
The output from the MCMC function is written to a binary file that can be read
back into R with the \code{read.rcm.dmm} function. For details on its structure
consult the documentation for \code{read.rcm.dmm}.
}
\description{
Normal-inverse Wishart Markov process model (random cluster model)
}
\seealso{
\code{\link{read.rcm.dmm}} for description of output file format.
}
