% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_dmm.R
\name{read.mk.dmm}
\alias{read.mk.dmm}
\title{Read the output file from \code{make.mk.dmm} into R}
\usage{
read.mk.dmm(output.file, skip = 0, n = -1)
}
\arguments{
\item{output.file}{Name of the output file.}

\item{skip}{Number of rows in the file to skip over before reading. Defaults to 0.}

\item{n}{Number of rows to read. If -1 (default) all rows (after any skipped
rows) are read until the end of file is reached.}
}
\value{
A list with the following structure
\describe{
\item{pars}{A numeric matrix whose rows correspond to posterior samples. The
first column contains the log likelihood of the data conferred by the specific
configuration of state assignments in the sample. The second column contains
the prior probability of the state configuration under the UCM model. The
third column contains the hyperparameter of the UCM prior model. The subsequent
columns are the hyperparameters of the Dirichlet prior on the multinomial
distribution of each state.}
\item{stateid}{An integer matrix that records the index of the state sampled
for each terminal taxon in each posterior sample. Note that the indices of
states in one sample do not necessarily align with the indices in another.}}
}
\description{
Read the output file from \code{make.mk.dmm} into R
}
\details{
The output file is a special binary file with the
following header format.
\tabular{rrl}{
Offset \tab Size  \tab  Description\cr
0      \tab 4     \tab  Number of resource categories (= J)\cr
4      \tab 4     \tab  Number of resource states (= K)\cr
8      \tab 4     \tab  Number of terminal taxa (= N)\cr
12     \tab 4     \tab  Number of rows in the input data matrix (= M)\cr
16     \tab 4     \tab  Number of characters in the Newick string for the input phylogeny\cr
20     \tab 6     \tab  The string "loglk"\cr
26     \tab 6     \tab  The string "prior"\cr
32     \tab 6     \tab  The string "alpha"\cr
38     \tab X     \tab  An array of strings giving the names of the resource categories\cr
38+X   \tab Y     \tab  An array of strings giving the names of the terminal taxa\cr
38+X+Y \tab 12M   \tab  An array of integers giving the input dataset\cr
38+X+Y+12M \tab Z \tab  The Newick string
}

Note that byte offsets assume 1 byte chars, 4 byte integers, and 8 byte doubles.
After the header is the payload. The first entry in the payload is the initial state
configuration and hyperparameter values that were used to start the MCMC run.
The payload has the following repetitive structure, where W initially equals
38+X+Y+12M+Z and is incremented by 24+8J+4N for each posterior sample.

\tabular{rrl}{
Offset  \tab   Size \tab  Description\cr
W       \tab   8    \tab  The data log likelihood\cr
W+8     \tab   8    \tab  The log prior probability\cr
W+16    \tab   8    \tab  The hyperparameter of the UCM prior\cr
W+24    \tab   8J   \tab  The hyperparameters of the Dirichlet prior\cr
W+24+8J \tab   4N   \tab  The state assignments of the terminal taxa
}
}
