% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_norm.R
\name{read.rcm.norm}
\alias{read.rcm.norm}
\title{Read the output file from \code{make.rcm.norm} into R}
\usage{
read.rcm.norm(output.file, skip = 0, n = -1)
}
\arguments{
\item{output.file}{Name of the output file.}

\item{skip}{Number of rows in the file to skip over before reading. Defaults to 0.}

\item{n}{Number of rows to read. If -1 (default) all rows (after any skipped
rows) are read until the end of file is reached.}
}
\value{
A list with the following structure
\describe{
\item{r}{The number of states in the prior model.}
\item{integrate.brlen}{Were branch lengths integrated out.}
\item{pars}{A numeric matrix whose rows correspond to posterior samples. The
first column contains the log likelihood of the data conferred by the specific
configuration of state assignments in the sample. The second column contains
the prior probability of the state configuration under the Markov model. The
third column contains the rate of the Markov model.}
\item{stateid}{An integer matrix that records the index of the state sampled
for each terminal taxon in each posterior sample. Note that the indices of
states in one sample do not necessarily align with the indices in another.}}
\item{kappa.init}{The hyperparameters of the Normal-inverse Wishart prior.}
\item{nu.init}{The hyperparameters of the Normal-inverse Wishart prior.}
\item{mu.init}{The hyperparameters of the Normal-inverse Wishart prior.}
\item{lambda.init}{The hyperparameters of the Normal-inverse Wishart prior.}
\item{dataset}{The dataset used for analysis. The first column holds the
terminal node indices, the second colum, the specimen labels,
the third column holds the character indices,
and the fourth column holds the measurment for the particular
combination of indices in columns one, two, and three. Note that the indices in the
first and third columns are 0-based, following C (not R) convention. Also, resource
category indices correspond to the name ordering in the \code{mu.init}
list member. Thus, index 0 corresponds to the first named element of
of \code{mu.init}; index 1, to the second; and so on.}
\item{phy}{The phylogeny used for analysis.}
}
\description{
Read the output file from \code{make.rcm.norm} into R
}
\details{
The output file is a special binary file with the
following header format.
\tabular{rrl}{
Offset \tab Size  \tab  Description\cr
0      \tab 4     \tab  Number of quantitative characters (= J)\cr
4      \tab 4     \tab  Number of resource states (= K)\cr
8      \tab 4     \tab  Flag indicating if branch lengths are integrated out\cr
12     \tab 4     \tab  Number of rows in the input data matrix (= M)\cr
16     \tab 4     \tab  Number of terminal taxa (= N)\cr
20     \tab 4     \tab  Number of characters in the Newick string for the input phylogeny\cr
24     \tab X     \tab  An array of strings giving the names of the terminal taxa\cr
24+X   \tab Y     \tab  An array of strings giving the names of the characters\cr
24+X+Y \tab 32M   \tab  An array of doubles giving the input dataset\cr
24+X+Y+32M \tab Z \tab  The Newick string\cr
24+X+Y+32M+Z \tab 8(2+J+J*J) \tab The hyperparameters of the Normal-inverse Wishart prior
}

Note that byte offsets assume 1 byte chars, 4 byte integers, and 8 byte doubles.
After the header is the payload. The first entry in the payload is the initial state
configuration and hyperparameter values that were used to start the MCMC run.
The payload has the following repetitive structure, where W initially equals
24+X+Y+32M+Z+8(2+J+J*J) and is incremented by 24+4N for each posterior sample.

\tabular{rrl}{
Offset  \tab   Size \tab  Description\cr
W       \tab   8    \tab  The data log likelihood\cr
W+8     \tab   8    \tab  The log prior probability\cr
W+16    \tab   8    \tab  The rate of the Poisson prior\cr
W+24    \tab   4N   \tab  The state assignments of the terminal taxa
}
}
