% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_epoch.R
\name{simulate.mk.epoch}
\alias{simulate.mk.epoch}
\title{Simulate node states under an Mk model}
\usage{
\method{simulate}{mk.epoch}(par, lik, n = 1L, root.p = NULL)
}
\arguments{
\item{par}{A vector of transition rate parameters for \code{lik}}

\item{lik}{A likelihood function returned by \code{make.mk.epoch}}

\item{n}{The number of simulations}

\item{root.p}{A vector of probabilities for drawing the root state.
The default is NULL}
}
\value{
A matrix of simulated node states for internal and terminal
nodes. Rows correspond to node indices and columns to simulation
number.
}
\description{
Simulate node states under an Mk model
}
\author{
Michael C. Grundler
}
