% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snakediet}
\alias{snakediet}
\title{Pseudoboine snake diets}
\format{
A dataframe with 91 rows and 3 columns
\describe{
\item{species}{Scientific name of the snake predator}
\item{food}{Type of prey}
\item{count}{Number of times the given prey has been recorded
from the diet of the particular species of snake}}
}
\source{
Alencar, L.R.V., Gaiarsa, M.P., and M. Martins. 2013. South American Journal of Herpetology 8(1): 60-66
}
\usage{
snakediet
}
\description{
A dataset from Alencar et al. (2013) containing counts of prey types
recovered from the stomachs of pseudobine snakes in South America.
}
\keyword{datasets}
