% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snaketree}
\alias{snaketree}
\title{Pseudoboine snake phylogeny}
\format{A Newick formatted character string}
\source{
Tonini, J., Beard, K.H., Ferreira, R.B., Jetz, W., and R.A. Pyron. 2016. Biological Conservation 204: 23-31
}
\usage{
snaketree
}
\description{
A phylogeny of evolutionary relationships among the pseudoboine snakes
included in the dataset from Alencar et al. (2013).
}
\keyword{datasets}
