% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroevolution.R
\docType{package}
\name{macroevolution}
\alias{macroevolution}
\alias{macroevolution-package}
\title{macroevolution: a package for comparative analysis of biological diversity}
\description{
macroevolution: a package for comparative analysis of biological diversity
}
\section{Tree I/O functions}{

read.newick
write.newick
}

\section{Tree info functions}{

ages
ancestors
brlens
children
descendants
Nnode
Ntip
root
tiplabels
tree.isbinary
tree.isultrametric
}

\section{Tree manipulation functions}{

extract.clade
keep.tip
rotate.node
tree.duplicate
tree.ladderize
}

\section{Tree visualization functions}{

plot.tree
}

\section{Functions for models of character evolution}{

make.mk.dmm
make.mk.dmm.from.sample
read.mk.dmm
}

